ref: extract pop up body
